Change Docker setup for Git workshop
In order for people to be able to run phpunit tests on any system, the
Dockerfile needed to be updated so that it includes installing composer
and (therefore) phpunit.
Also sneaked in a fix for stupid segmentation errors in tests we don't
care about for the purposes of this fork.